ET with millisecond precision

See merge request ecm/ECM!804

+++ source/Lib/EncoderLib/EncGOP.h
@@ -274,7 +274,7 @@ protected:
 #endif
 
   void  xCalculateAddPSNRs(const bool isField, const bool isFieldTopFieldFirst, const int iGOPid, Picture* pcPic, 
-    const AccessUnit&accessUnit, PicList &rcListPic, int64_t dEncTime, const InputColourSpaceConversion snr_conversion, 
+    const AccessUnit&accessUnit, PicList &rcListPic, double dEncTime, const InputColourSpaceConversion snr_conversion,
     const bool printFrameMSE, const bool printMSSSIM, double* PSNR_Y, bool isEncodeLtRef);
   void  xCalculateAddPSNR(Picture* pcPic, PelUnitBuf cPicD, const AccessUnit&, double dEncTime, const InputColourSpaceConversion snr_conversion, 
     const bool printFrameMSE, const bool printMSSSIM, double* PSNR_Y, bool isEncodeLtRef);
